enemy death count, bug fixes

+++ WitchPack/Assets/Scripts/Gameplay/Units/UnitStats.cs
@@ -22,7 +22,7 @@ public class UnitStats
     private int baseDamage;
     private float attackSpeed;
     private int baseRange;
-    private int movementSpeed;
+    private float movementSpeed;
     private int critDamage;
     private int critChance;
     private float invincibleTime;
@@ -37,7 +37,7 @@ public class UnitStats
     public int BaseDamage { get { return Mathf.RoundToInt(Mathf.Clamp((ownerBaseStats.BaseDamage.value + baseDamage), 0, (ownerBaseStats.BaseDamage.value + baseDamage))); } }
     public float AttackSpeed { get { return Mathf.Clamp((ownerBaseStats.AttackSpeed.value + attackSpeed), 0, 2f); } }
     public int BonusRange { get { return Mathf.RoundToInt(Mathf.Clamp((ownerBaseStats.BaseRange.value + baseRange), 0, (ownerBaseStats.BaseRange.value + baseRange))); } }
-    public int MovementSpeed { get { return Mathf.RoundToInt(Mathf.Clamp((ownerBaseStats.MovementSpeed.value + movementSpeed), 1, (ownerBaseStats.MovementSpeed.value + movementSpeed))); } }
+    public float MovementSpeed { get { return Mathf.RoundToInt(Mathf.Clamp((ownerBaseStats.MovementSpeed.value + movementSpeed), 1, (ownerBaseStats.MovementSpeed.value + movementSpeed))); } }
     public int CritDamage { get { return Mathf.RoundToInt(Mathf.Clamp((ownerBaseStats.CritDamage.value + critDamage), 0, (ownerBaseStats.CritDamage.value + critDamage))); } }
     public int CritChance { get { return Mathf.RoundToInt(Mathf.Clamp((ownerBaseStats.CritChance.value + critChance), 0, (ownerBaseStats.CritChance.value + critChance))); } }
     public float InvincibleTime { get { return Mathf.Clamp((ownerBaseStats.InvincibleTime.value + invincibleTime), 0, 0.5f); } }
